[SDCISA-13746] Delegate redis hset call to worker thread
For whatever reason hset seems to take a while. Not yet sure what is
going on. According to the stacktraces of the BlockedThreadChecker it
seems as BLOCKING IO is happening inside hset. But instead, it could
also be that caller itself gets called from some forEach loops, and
maybe the iterated thingy is too large to do everything in one eventloop
task. Either way, lets move that operation over to a worker thread and
lets see if it changes anything.

(wip @ eb3af5aac754bb17b781127be28cfb9b246ed96d)